refactor: move session logic to new folder
