+++ packages/core/src/sessions/SessionStorage.ts
@@ -16,30 +16,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {useCallback, useEffect, useMemo, useState} from 'react';
 import {storage, Storage} from '../storage/Storage';
+import {Session} from './type';
+import {setActiveSession} from './utils';
 
 const SESSION_KEY = 'ConnectionSessions';
-
-const setActiveSession = (
-  sessionList: Session[],
-  activeSessionId: string,
-): Session[] => {
-  return sessionList.map(_session => {
-    if (_session.id === activeSessionId) {
-      return {..._session, isActive: true};
-    }
-
-    return {..._session, isActive: false};
-  });
-};
-
-interface Session {
-  id: string;
-  url: string;
-  username: string;
-  isActive: boolean;
-}
 
 class SessionStorage {
   private refreshCallBack: ({
@@ -128,36 +109,3 @@ class SessionStorage {
 }
 
 export const sessionStorage = new SessionStorage(storage);
-
-export const useSessions = (enableConnectionSessions: boolean = false) => {
-  const [sessionList, setSessionList] = useState(
-    sessionStorage.getSessionList(),
-  );
-
-  const [sessionActive, setSessionActive] = useState(
-    sessionStorage.getActiveSession(),
-  );
-
-  const refreshData = useCallback(({sessionList: sessions, activeSession}) => {
-    setSessionActive(activeSession);
-    setSessionList(sessions);
-  }, []);
-
-  useEffect(() => {
-    sessionStorage.register(refreshData);
-  }, [refreshData]);
-
-  return useMemo(() => {
-    if (enableConnectionSessions) {
-      return {
-        sessionList: sessionList,
-        sessionActive: sessionActive,
-      };
-    }
-
-    return {
-      sessionList: [],
-      sessionActive: null,
-    };
-  }, [enableConnectionSessions, sessionActive, sessionList]);
-};
