+++ packages/core/src/sessions/type.ts
@@ -1,0 +1,6 @@
+export interface Session {
+  id: string;
+  url: string;
+  username: string;
+  isActive: boolean;
+}
